add project_tags for space_tags
