--- tap_teamwork/streams/tags.py
@@ -1,32 +1,0 @@
-from typing import List, Optional, Dict, Any
-from singer import get_logger
-from tap_teamwork.streams.abstracts import FullTableStream
-
-LOGGER = get_logger()
-
-
-class Tags(FullTableStream):
-    tap_stream_id = "tags"
-    parent = "spaces"
-    key_properties = ["id"]
-    replication_method = "FULL_TABLE"
-    replication_keys: List[str] = []
-    data_key = "tags"
-
-    def get_url_endpoint(self, parent_obj: Optional[Dict[str, Any]] = None) -> str:
-        if not parent_obj:
-            raise ValueError("Missing parent_obj for tags stream")
-
-        space_id = parent_obj.get("id")
-        if not space_id:
-            raise ValueError("Missing 'id' in parent_obj for tags stream")
-
-        LOGGER.info("Fetching tags for id=%s", space_id)
-        return self.client.build_url("spaces/api/v1/tags.json")
-
-    def get_child_context(
-        self,
-        record: Dict[str, Any],
-        context: Optional[Dict[str, Any]],
-    ) -> Optional[Dict[str, Any]]:
-        return None
